% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_cells.R
\name{aggregate_cells}
\alias{aggregate_cells}
\title{Aggregate cells}
\usage{
aggregate_cells(.data, .sample)
}
\arguments{
\item{.data}{}

\item{.sample}{}
}
\value{
Object of same type as .data
}
\description{
Aggregate cells
}
