% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mito_info_all_datasets}
\alias{mito_info_all_datasets}
\title{mito_info_all_datasets}
\format{
A Seurat object.
}
\usage{
data(mito_info_all_datasets)
}
\description{
A SingleCellExperiment dataset of single cell RNA sequencing data. Used for one of the exercises
}
\keyword{datasets}
