% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{pbmc}
\format{
A SingleCellExperiment object.
}
\usage{
data(pbmc)
}
\description{
A SingleCellExperiment dataset of single cell RNA sequencing data
}
\keyword{datasets}
