% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_seurat}
\alias{pbmc_seurat}
\title{a subset version of pbmc in seurat format}
\format{
A Seurat object.
}
\usage{
data(pbmc_seurat)
}
\description{
A Seurat dataset of single cell RNA sequencing data
}
\keyword{datasets}
